requirements.txt
python-dotenv~=1.0.1
pydantic~=2.10.6
Telethon~=1.39.0
sqlalchemy~=2.0.38
fastapi~=0.115.8
uvicorn~=0.32.1
asyncpg
alembic